Fix OverflowException for server port
Change short to int. Bug report by 1092CQ.

+++ MinecraftClient/McTcpClient.cs
@@ -53,7 +53,7 @@ namespace MinecraftClient
         /// <param name="port">The server port to use</param>
         /// <param name="protocolversion">Minecraft protocol version to use</param>
 
-        public McTcpClient(string username, string uuid, string sessionID, int protocolversion, string server_ip, short port)
+        public McTcpClient(string username, string uuid, string sessionID, int protocolversion, string server_ip, int port)
         {
             StartClient(username, uuid, sessionID, server_ip, port, protocolversion, false, "");
         }
@@ -69,7 +69,7 @@ namespace MinecraftClient
         /// <param name="protocolversion">Minecraft protocol version to use</param>
         /// <param name="command">The text or command to send.</param>
 
-        public McTcpClient(string username, string uuid, string sessionID, string server_ip, short port, int protocolversion, string command)
+        public McTcpClient(string username, string uuid, string sessionID, string server_ip, int port, int protocolversion, string command)
         {
             StartClient(username, uuid, sessionID, server_ip, port, protocolversion, true, command);
         }
@@ -86,7 +86,7 @@ namespace MinecraftClient
         /// <param name="singlecommand">If set to true, the client will send a single command and then disconnect from the server</param>
         /// <param name="command">The text or command to send. Will only be sent if singlecommand is set to true.</param>
 
-        private void StartClient(string user, string uuid, string sessionID, string server_ip, short port, int protocolversion, bool singlecommand, string command)
+        private void StartClient(string user, string uuid, string sessionID, string server_ip, int port, int protocolversion, bool singlecommand, string command)
         {
             this.sessionid = sessionID;
             this.uuid = uuid;
